More python work

+++ pet-race-job/pet_race_job/model/pet_categories.py
@@ -15,7 +15,7 @@ from cassandra.cqlengine.models import Model
 class PetCategories(Model):
     petCategoryId = columns.UUID(primary_key=True, default=uuid.uuid4)
     name = columns.Text(required=True, index=True)
-    speed = columns.BigInt(required=False)
+    speed = columns.Float(required=False)
 
 
 __keyspace__ = 'gpmr'
